pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.18.0",
    "aiopg>=1.4.0",
    "loguru>=0.7.3",
    "mixpanel>=4.10.1",
    "peewee>=3.17.9",
    "peewee-async>=1.0.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
]
